Fix bugs

+++ MessagingSystem.Services.User/Application/User/UserOrchestrator.cs
@@ -1,5 +1,7 @@
 using AutoMapper;
 using FluentValidation;
+using MassTransit;
+using MessagingSystem.SendingModels.UserNotification;
 using MessagingSystem.Services.User.Core.User;
 using MessagingSystem.Services.User.Infrastructure.Encrypt;
 using MessagingSystem.Services.User.Infrastructure.HasherInfo;
@@ -11,7 +13,8 @@ public class UserOrchestrator(
     IValidator<UserDto> validator, 
     IUserRepository userRepository,
     IEncryptInfo encryptInfo,
-    IHasher hasher)
+    IHasher hasher,
+    IPublishEndpoint publishEndpoint)
 {
     public async Task<OperationResult<string>> EditUserInfoAsync(UserDto userDto, string userId)
     {
@@ -27,17 +30,24 @@ public class UserOrchestrator(
         var hashLogin = hasher.Hash(userDto.Login);
         var hashEmail = hasher.Hash(userDto.Email);
         var hashNickName = hasher.Hash(userDto.NickName);
-    
-        if (encryptInfo is EncryptInfo concreteEncryptor)
-            concreteEncryptor.EncryptObjectStrings(userDto);
-
+        
         mapper.Map(userDto, existingUser);
-
         existingUser.SetHashes(hashLogin, hashEmail, hashNickName);
-
+        
+        if (encryptInfo is EncryptInfo concreteEncryptor)
+            concreteEncryptor.EncryptObjectStringsForUpdate(existingUser);
+        
+        Console.WriteLine(existingUser.UserName);
         try
         {
+            if (existingUser.UserName == null || existingUser.Email == null) 
+                return OperationResult<string>.Fail("User can not have null properties");
+            
             await userRepository.UpdateUserAsync(existingUser);
+            
+            var editUserInfo = new EditUserEmail(existingUser.Email, existingUser.UserName);
+            await publishEndpoint.Publish(editUserInfo);
+            
             return OperationResult<string>.Ok("Update user info");
         }
         catch (Exception e)
@@ -55,7 +65,14 @@ public class UserOrchestrator(
         
         try
         {
+            if (user.UserName == null || user.Email == null) 
+                return OperationResult<string>.Fail("User can not have null properties");
+            
             await userRepository.DeleteUserAsync(user);
+            
+            var editUserInfo = new DeleteUserEmail(user.Email, user.UserName);
+            await publishEndpoint.Publish(editUserInfo);
+            
             return OperationResult<string>.Ok("Delete user");
         }
         catch (Exception e)
